entity , dto, repositoryImpl --> overuse_policy - unit_price(int --> double)로 수정

+++ billing_batch/src/main/java/com/touplus/billing_batch/domain/dto/OverusePolicyDto.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class OverusePolicyDto {
     private Long overusePolicyId;
     private UseType useType;
-    private Integer unitPrice;
+    private Double unitPrice;
 
     /**
      * Entity -> DTO 변환
